+++ NEWS.md
@@ -6,6 +6,8 @@
   - Introduction
   - Parameters
   - RStudio Addin
+- add abnormal volume R6 class
+- add abnormal volatility R6 class
 
 # Version 0.1.1